refactor: Move shouldDisplaySection into utils folder
