+++ packages/ui-app/src/InputExtrinsic/options/section.ts
@@ -5,23 +5,9 @@
 import { Extrinsic$Sections } from '@polkadot/extrinsics/types';
 
 import { DropdownOptions } from '../types';
+import { shouldDisplaySection } from '../../util/shouldDisplaySection';
 
 import map from '@polkadot/extrinsics';
-
-export function anyMethodToDisplay (methods: any) {
-  return Object
-    .keys(methods)
-    .filter((name) => {
-      const { isDeprecated, isHidden } = methods[name];
-
-      return !isDeprecated && !isHidden;
-    })
-    .length !== 0;
-}
-
-export function shouldDisplaySection (section: any, methods: any) {
-  return !section.isDeprecated && !section.isHidden && anyMethodToDisplay(methods);
-}
 
 export default function createOptions (type: 'private' | 'public'): DropdownOptions {
   return Object
