+++ packages/ui-app/src/util/shouldDisplaySection.spec.js
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the ISC license. See the LICENSE file for details.
 
-import { shouldDisplaySection } from './section';
+import { shouldDisplaySection } from './shouldDisplaySection';
 
 describe('input extrinsic section', () => {
   let type, dummySection;
